- Enable -Wall flag in CMakeLists
- Address compiler warnings such as mixed-sign comparison, initialization order, unused variables

+++ includes/Algo/DynamicRange/AsymClip.h
@@ -3,12 +3,12 @@
 /*
  * /////////
  * // Clover
- * 
+ *
  * Audio processing algorithms and DAG with feedback loops that do not break
  * acyclicity.
- * 
+ *
  * Copyright (C) 2023 Rob W. Albus
- *  
+ *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any later
@@ -17,9 +17,8 @@
  * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
  * details. You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
- * 
+ *
  */
-
 
 #include <array>
 
@@ -36,7 +35,7 @@ struct AsymClip : public AlgorithmBase<std::array<T, __arity>> {
       : minimum_(minimum), maximum_(maximum) {}
 
   std::array<T, __arity> process(std::array<T, __arity> input) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       T sign = Util::Calc::sign(input[i]);
       if (sign < 0) {
         this->processed[i] = std::max(minimum_, input[i]);
