+++ includes/Algo/DynamicRange/UpwardDCCompression.h
@@ -3,12 +3,12 @@
 /*
  * /////////
  * // Clover
- * 
+ *
  * Audio processing algorithms and DAG with feedback loops that do not break
  * acyclicity.
- * 
+ *
  * Copyright (C) 2023 Rob W. Albus
- *  
+ *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any later
@@ -17,9 +17,8 @@
  * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
  * details. You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
- * 
+ *
  */
-
 
 #include <array>
 #include <cmath>
@@ -36,7 +35,7 @@ struct UpwardDCCompression : public AlgorithmBase<std::array<T, __arity>> {
 
   std::array<T, __arity> process(std::array<T, __arity> input) {
     std::array<T, __arity> frame{};
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       T sign = input[i] < T(0) ? T(-1) : T(1);
       T offset = threshold_ * sign;
       T result = input[i] * scaleAmount_ + offset;
