+++ includes/Graph/AudioFrame.h
@@ -3,12 +3,12 @@
 /*
  * /////////
  * // Clover
- * 
+ *
  * Audio processing algorithms and DAG with feedback loops that do not break
  * acyclicity.
- * 
+ *
  * Copyright (C) 2023 Rob W. Albus
- *  
+ *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any later
@@ -17,9 +17,8 @@
  * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
  * details. You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
- * 
+ *
  */
-
 
 #include <array>
 #include <cstdio>
@@ -49,7 +48,7 @@ template <size_t __arity> struct AudioFrame {
   Graph::AudioFrame<__arity>
   operator+(const Graph::AudioFrame<__arity> &b) const {
     Graph::AudioFrame<__arity> f;
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       f[i] = data[i] + b[i];
     }
     return f;
@@ -58,7 +57,7 @@ template <size_t __arity> struct AudioFrame {
   Graph::AudioFrame<__arity>
   operator-(const Graph::AudioFrame<__arity> &b) const {
     Graph::AudioFrame<__arity> f;
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       f[i] = data[i] - b[i];
     }
     return f;
@@ -67,7 +66,7 @@ template <size_t __arity> struct AudioFrame {
   Graph::AudioFrame<__arity>
   operator*(const Graph::AudioFrame<__arity> &b) const {
     Graph::AudioFrame<__arity> f;
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       f[i] = data[i] * b[i];
     }
     return f;
@@ -76,62 +75,62 @@ template <size_t __arity> struct AudioFrame {
   Graph::AudioFrame<__arity>
   operator/(const Graph::AudioFrame<__arity> &b) const {
     Graph::AudioFrame<__arity> f;
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       f[i] = data[i] / b[i];
     }
     return f;
   }
 
   void operator+=(const Graph::AudioFrame<__arity> &b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] += b[i];
     }
   }
 
   void operator-=(const Graph::AudioFrame<__arity> &b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] -= b[i];
     }
   }
 
   void operator*=(const Graph::AudioFrame<__arity> &b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] *= b[i];
     }
   }
 
   void operator/=(const Graph::AudioFrame<__arity> &b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] /= b[i];
     }
   }
 
   void operator+=(const float b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] += b;
     }
   }
 
   void operator-=(const float b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] -= b;
     }
   }
 
   void operator*=(const float b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] *= b;
     }
   }
 
   void operator/=(const float b) {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] /= b;
     }
   }
 
   void init() {
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       data[i] = 0;
     }
   }
