+++ includes/NodeSimplex/Deprecate/RingMod_V1.h
@@ -3,12 +3,12 @@
 /*
  * /////////
  * // Clover
- * 
+ *
  * Audio processing algorithms and DAG with feedback loops that do not break
  * acyclicity.
- * 
+ *
  * Copyright (C) 2023 Rob W. Albus
- *  
+ *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any later
@@ -17,9 +17,8 @@
  * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
  * details. You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
- * 
+ *
  */
-
 
 #include <cmath>
 
@@ -40,7 +39,7 @@ private:
   Graph::AudioFrame<__arity> tick(Graph::AudioFrame<__arity> input) {
     float shape = ((_shape * 0.5) + 0.5) * M_PI;
     Graph::AudioFrame<__arity> f{};
-    for (int i = 0; i < __arity; i++) {
+    for (size_t i = 0; i < __arity; i++) {
       f[i] = sin(shape * input[i]);
     }
